pyproject.toml
[project]
name = "pcb2blender"
version = "2.17.0"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = ["error-helper==1.4", "pillow==11.1.0", "skia-python==87.6"]
license = "GPL-3.0-or-later"

[dependency-groups]
dev = [
    "fake-bpy-module",
    "numpy",
    "pip",
    "pytest",
    "pytest-blender",
    "requests",
    "snakeviz",
    "tomlkit",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint.isort]
combine-as-imports = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "blender",
    "pcbnew",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
blender = ["addon_utils", "bpy", "bpy_extras", "bl_ui", "bmesh", "mathutils", "numpy"]
pcbnew = ["pcbnew", "wx"]
